Refactor IPC
